requirements.txt
robotframework==7.2.2
paramiko>=2.5.0
requests
robotframework-requests
robotframework-jsonlibrary
robotframework-sshlibrary
robotframework-scplibrary
robotframework-httplibrary
redfish>=3.1.7
redfishtool>=1.1.8
